ai-smart-assistant/requirements.txt
streamlit
requests
python-dotenv
duckduckgo-search
PyPDF2
bs4
langchain-community
tiktoken
sentence-transformers
faiss-cpu
certifi